+++ js/utils/open-window.js
@@ -1,10 +1,7 @@
 
-const openWindow = (openButton, activeWindow) => {
-  openButton.addEventListener('change', (evt) => {
-    evt.preventDefault();
-    activeWindow.classList.remove('hidden');
-    document.querySelector('body').classList.remove('.modal-open');
-  });
+const openWindow = (activeWindow) => {
+  activeWindow.classList.remove('hidden');
+  document.querySelector('body').classList.remove('.modal-open');
 };
 
 export {openWindow};
